# api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "Category" : {
        "properties" : {
          "code" : {
            "example" : "GROCERIES",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "description" : "Only SPEND counts towards spent",
            "enum" : [ "SPEND", "INCOME", "TRANSFER" ],
            "type" : "string"
          },
          "name" : {
            "example" : "Groceries",
            "type" : "string"
          },
          "parentId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "code", "id", "kind", "name" ],
        "type" : "object"
      },
...
      "CategoryDetail" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "averageMinor" : {
            "description" : "Average over the trend months that have any data; absent when none has",
            "format" : "int64",
            "type" : "integer"
          },
          "category" : {
            "$ref" : "#/components/schemas/Category"
          },
          "currency" : {
            "example" : "RON",
            "type" : "string"
          },
          "merchants" : {
            "description" : "This month's variable spending by merchant (recurring charges excluded), largest first",
            "items" : {
              "$ref" : "#/components/schemas/MerchantAmount"
            },
            "type" : "array"
          },
          "month" : {
            "example" : "2026-03",
            "type" : "string"
          },
          "recurring" : {
            "description" : "This month's recurring payments, largest first",
            "items" : {
              "$ref" : "#/components/schemas/RecurringAmount"
            },
            "type" : "array"
          },
          "recurringMinor" : {
            "description" : "This month's recurring charges, total",
            "format" : "int64",
            "type" : "integer"
          },
          "trend" : {
            "description" : "The 12 months ending with `month`, oldest first",
            "items" : {
              "$ref" : "#/components/schemas/MonthAmount"
            },
            "type" : "array"
          }
        },
        "required" : [ "amountMinor", "category", "currency", "merchants", "month", "recurring", "recurringMinor", "trend" ],
        "type" : "object"
      },
      "CategorySpend" : {
        "properties" : {
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "code" : {
            "type" : "string"
          },
          "deltaPct" : {
            "description" : "Per month vs. usual; absent when there is no usual to compare with",
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "sharePct" : {
            "description" : "Of the month's spent",
            "format" : "int32",
            "type" : "integer"
          },
          "spentMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "transactionCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "usualMinor" : {
            "description" : "Per-month average over the baseline months (the 3 before the period)",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "name", "sharePct", "spentMinor", "transactionCount", "usualMinor" ],
        "type" : "object"
      },
      "Committed" : {
        "properties" : {
          "count" : {
            "format" : "int32",
            "type" : "integer"
          },
          "incomeCount" : {
            "description" : "Recurring income streams counted (a salary in two parts is two)",
            "format" : "int32",
            "type" : "integer"
          },
          "incomeMonthlyMinor" : {
            "description" : "Confirmed recurring income per month, as on the Recurring screen for this month",
            "format" : "int64",
            "type" : "integer"
          },
          "monthlyMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "sharePct" : {
            "description" : "Of the month's spent; absent when nothing was spent",
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "count", "incomeCount", "incomeMonthlyMinor", "monthlyMinor" ],
        "type" : "object"
      },
...
      "Insight" : {
        "properties" : {
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "code" : {
            "type" : "string"
          },
          "deltaPct" : {
            "description" : "vs. usual, e.g. 40 or -59",
            "format" : "int32",
            "type" : "integer"
          },
          "differenceMinor" : {
            "description" : "Per month minus usual; negative = less",
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "perMonthMinor" : {
            "description" : "Spent per month in the period",
            "format" : "int64",
            "type" : "integer"
          },
          "transactionCount" : {
            "description" : "Transactions in the period (visits)",
            "format" : "int32",
            "type" : "integer"
          },
          "usualMinor" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "categoryId", "deltaPct", "differenceMinor", "name", "perMonthMinor", "transactionCount", "usualMinor" ],
        "type" : "object"
      },
...
      "MerchantAmount" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "merchantId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "transactionCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "amountMinor", "merchantId", "name", "transactionCount" ],
        "type" : "object"
      },
...
      "MonthAmount" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "hasData" : {
            "description" : "Whether any transaction exists in this month",
            "type" : "boolean"
          },
          "month" : {
            "example" : "2026-03",
            "type" : "string"
          }
        },
        "required" : [ "amountMinor", "hasData", "month" ],
        "type" : "object"
      },
      "MonthSummary" : {
        "properties" : {
          "accuracyPct" : {
            "description" : "Share of spending weighted by category confidence (before reviewedPct; kept for reference)",
            "format" : "int32",
            "type" : "integer"
          },
          "availableMonths" : {
            "description" : "Months with data, oldest first, e.g. 2026-01",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "averageSpentMinor" : {
            "description" : "Average spent over the baseline months; absent without history",
            "format" : "int64",
            "type" : "integer"
          },
          "baselineMonths" : {
            "description" : "How many of the previous 3 months have data; the average uses only those",
            "format" : "int32",
            "type" : "integer"
          },
          "categories" : {
            "description" : "Top 5 by spent, largest first",
            "items" : {
              "$ref" : "#/components/schemas/CategorySpend"
            },
            "type" : "array"
          },
          "categorizedPct" : {
            "description" : "Share of this month's spending with any category",
            "format" : "int32",
            "type" : "integer"
          },
          "committed" : {
            "$ref" : "#/components/schemas/Committed",
            "description" : "Block 3: confirmed recurring payments"
          },
          "currencies" : {
            "description" : "Currencies with money in the selected accounts, most used first; more than one = a switch",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "currency" : {
            "example" : "RON",
            "type" : "string"
          },
          "deltaPct" : {
            "description" : "Spent vs. that average, e.g. 12 for +12%; absent without history or when it is 0",
            "format" : "int32",
            "type" : "integer"
          },
          "incomeMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "insight" : {
            "$ref" : "#/components/schemas/Insight",
            "description" : "The plain-language insight line; absent when nothing moved notably"
          },
          "month" : {
            "example" : "2026-03",
            "type" : "string"
          },
          "months" : {
            "description" : "1, or 3 for the 'last 3 months' view: figures are totals over the months ending with `month`",
            "format" : "int32",
            "type" : "integer"
          },
          "monthsWithData" : {
            "description" : "Months of the period that have data: totals ÷ this = per-month averages",
            "format" : "int32",
            "type" : "integer"
          },
          "netMinor" : {
            "description" : "Income − spent; negative when spending more",
            "format" : "int64",
            "type" : "integer"
          },
          "periodFrom" : {
            "description" : "First month of the period (YYYY-MM)",
            "type" : "string"
          },
          "rest" : {
            "$ref" : "#/components/schemas/Rest",
            "description" : "Everything below the top 5, folded"
          },
          "reviewedPct" : {
            "description" : "DESIGN's accuracy: share of the period's spending categorized and reviewed (Scope.REVIEWED)",
            "format" : "int32",
            "type" : "integer"
          },
          "spentMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "uncategorizedCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "uncategorizedMinor" : {
            "description" : "Spent without a category, wherever it ranks (it may be folded into the rest)",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "accuracyPct", "availableMonths", "baselineMonths", "categories", "categorizedPct", "committed", "currencies", "currency", "incomeMinor", "month", "months", "monthsWithData", "netMinor", "periodFrom", "rest", "reviewedPct", "spentMinor", "uncategorizedCount", "uncategorizedMinor" ],
        "type" : "object"
      },
...
      "RecurringAmount" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "cadence" : {
            "enum" : [ "DAILY", "WEEKLY", "BIWEEKLY", "MONTHLY", "QUARTERLY", "YEARLY" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "state" : {
            "description" : "CONFIRMED or ENDED",
            "type" : "string"
          },
          "subscriptionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "transactionCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "amountMinor", "cadence", "name", "state", "subscriptionId", "transactionCount" ],
        "type" : "object"
      },
...
      "Rest" : {
        "properties" : {
          "categoryCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sharePct" : {
            "format" : "int32",
            "type" : "integer"
          },
          "spentMinor" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "categoryCount", "sharePct", "spentMinor" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "Outflow API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/insights/categories/{id}" : {
      "get" : {
        "operationId" : "category",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "YYYY-MM",
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "ISO currency; absent = the main currency of the accounts (most transactions)",
          "in" : "query",
          "name" : "currency",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Account ids to include (accounts filter); absent = all accounts",
          "in" : "query",
          "name" : "accounts",
          "required" : false,
          "schema" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "insights" ]
      }
    },
    "/api/insights/month" : {
      "get" : {
        "operationId" : "month",
        "parameters" : [ {
          "description" : "YYYY-MM; defaults to the latest month with data",
          "in" : "query",
          "name" : "month",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "ISO currency; absent = the main currency of the accounts (most transactions)",
          "in" : "query",
          "name" : "currency",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "1, or 3 for the 'last 3 months' view ending with the month",
          "in" : "query",
          "name" : "months",
          "required" : false,
          "schema" : {
            "default" : 1,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "description" : "Account ids to include (accounts filter); absent = all accounts",
          "in" : "query",
          "name" : "accounts",
          "required" : false,
          "schema" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MonthSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "insights" ]
      }
    },
...
  },
  "servers" : [ {
    "url" : "/"
  } ]
}